Use integer-valued hashing keys for bindings, so as to create sets of solution groups

+++ sesamestream-impl/src/test/java/edu/rpi/twc/sesamestream/tuple/BindingsTest.java
@@ -22,8 +22,8 @@ public class BindingsTest {
         Query.QueryVariables vars
                 = new Query.QueryVariables(Arrays.asList("x", "y", "z"));
         Map<String, String> map = new HashMap<String, String>();
-        Set<Long> hashes = new HashSet<Long>();
-        long hash;
+        Set<Integer> hashes = new HashSet<Integer>();
+        int hash;
 
         map.clear();
         map.put("x", "red");
